requirements.txt
awesome-slugify==1.6.5
boto3==1.17.33
botocore==1.20.33
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
idna==2.10
jmespath==0.10.0
pymongo==3.11.3
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
regex==2021.3.17
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
Unidecode==0.4.21
urllib3==1.26.4
